[refactor] 기존의 모달 컴포넌트를 따로 분류
- 기존의 모달 컴포넌트를 따로 분류
- 절대경로 추가
